Refs #296: only set timezone if we have the setting.
Every release of PHP seems to try harder and harder to force folks to
set date.timezone in their php.ini files. date_default_timezone_get now
no longer checks the TZ environment variable and also returns UTC if one
is not set. This makes it less useful to use in any context.

I think it's time to stop fighting the tide and just set the TZ if
requested via the setting.