Add a class loader to map other dirs under the zenmagick ns
This is just the initial draft that prepares us for taking advantage
of the new composer classmap generation code when it lands. It is meant
to supplement the existing composer classloader during this
transistionary period.

It also makes clear some of the assumptions we make about directory
structures and gives us a chance to remedy them.

It supports all the current app name spaces (hardcoded), plugins,
,themes and shared.

I expect that we will remove the lib directories under apps, plugins,
and themes which would really simplify the implementation.