rely completely on the swiftmailer bundle for mail settings
This means we use swiftmailer prefixed container parameters to configure
email.

This rearranges the logic a bit in that we directly configure
swiftmailer in app/store/config/email.php via the ZC configuration table
parameters for store,storefront, and admin. Other apps can be configured
either in their own container file or by using the imports: key to
choose the file yourself.