Use single front controller for storefront and admin.
This refactors the routing loader to be more symfony like. Now we use a
single front controller for admin and storefront via the prefix /admin

Currently we have to check REQUEST_URI to tell the difference, but i
think that hack is worth it to move on to the next step in removing
context altogether.