use sf controller naming pattern for bundles
This is the default way symfony controllers are referenced in routes.

An example:
  AdminBundle:Login:process is equivalent to
  ZenMagick\AdminBundle\Controller\LoginController::processAction

This was easy to achieve by adding a processAction method to
ZMController. It's certainly not the final iteration of what we want to
do there, but it is much easier to type and conforms to existing
standards.